Changing the interval to decimal in interval time trigger
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C24798

+++ Vanrise/Code/Vanrise.Runtime.Business/TimeSchedulerTaskTrigger/IntervalTimeSchedulerTaskTrigger.cs
@@ -10,7 +10,7 @@ namespace Vanrise.Runtime.Business
 
     public class IntervalTimeSchedulerTaskTrigger : TimeSchedulerTaskTrigger
     {
-        public int Interval { get; set; }
+        public double Interval { get; set; }
 
         public IntervalType IntervalType { get; set; }
